+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## 25.10.4
 
-- Sending of HTML clipboard content was fixed (for [#3371](https://github.com/pbek/QOwnNotes/issues/3371))
+- There now is a menu action to the **send clipboard as text** to another instance
+  of QOwnNotes (for [#3371](https://github.com/pbek/QOwnNotes/issues/3371))
+- **Sending of HTML clipboard content** to another instance of QOwnNotes was fixed
+  (for [#3371](https://github.com/pbek/QOwnNotes/issues/3371))
 - A PPA for Ubuntu 26.04 (Resolute Raccoon) was added
   - See [Install on Ubuntu](https://www.qownnotes.org/installation/ubuntu.html)
     for more information on how to install the PPA